Move set_default_page_edit_handlers and _get_page_edit_handler to wagtail.admin.panels.page_utils
